+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Enhanced reset-documentation ability to delete all child terms and reset sync metadata
+- Complete codebase to project nomenclature cleanup
 
 ## [0.9.1] - 2026-01-24