+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Change sync to target depth-0 project terms
+
 ## [0.9.3] - 2026-01-26
 
 - Fix delete_child_terms() to filter by parent in PHP (parent__not_in not valid)